+++ src/common/register-clients.ts
@@ -13,6 +13,7 @@ import { CollaboratorClient } from "../collaborator/client";
 import { FunctionalTestingClient } from "../functional-testing/client";
 import { PactflowClient } from "../pactflow/client";
 import { QmetryClient } from "../qmetry/client";
+import { Qtm4jClient } from "../qtm4j/client";
 import { ReflectClient } from "../reflect/client";
 import { SwaggerClient } from "../swagger/client";
 import { ZephyrClient } from "../zephyr/client";
@@ -35,6 +36,9 @@ clientRegistry.register(new QmetryClient());
 
 // Register Zephyr client
 clientRegistry.register(new ZephyrClient());
+
+// Register QTM4J client
+clientRegistry.register(new Qtm4jClient());
 
 // Register Collaborator client
 clientRegistry.register(new CollaboratorClient());